+++ lib/td_dd/application.ex
@@ -18,6 +18,11 @@ defmodule TdDd.Application do
       worker(TdDd.Search.IndexWorker, [TdDd.Search.IndexWorker]),
       # Worker for background bulk loading
       worker(TdDd.Loader.LoaderWorker, [TdDd.Loader.LoaderWorker]),
+      # Workers for cache loading
+      worker(TdDd.Cache.SystemLoader, []),
+      worker(TdDd.Cache.StructureLoader, []),
+      worker(TdDd.Cache.FieldLoader, []),
+      # Duplicate removal task on startup
       {TdDd.DataStructures.DuplicateRemover, []}
     ]