+++ lib/td_dd_web/controllers/echo_controller.ex
@@ -1,9 +1,11 @@
 defmodule TdDdWeb.EchoController do
   use TdDdWeb, [:controller, :warn]
 
-  action_fallback TdDdWeb.FallbackController
+  alias Jason, as: JSON
+
+  action_fallback(TdDdWeb.FallbackController)
 
   def echo(conn, params) do
-    send_resp(conn, 200, params |> Poison.encode!)
+    send_resp(conn, 200, params |> JSON.encode!())
   end
 end