+++ lib/td_dd_web/controllers/metadata_controller.ex
@@ -1,15 +1,16 @@
 defmodule TdDdWeb.MetadataController do
-  require Logger
   use TdDdWeb, :controller
 
+  alias Jason, as: JSON
   alias Plug.Upload
+  alias TdCache.TaxonomyCache
   alias TdDd.Auth.Guardian.Plug, as: GuardianPlug
   alias TdDd.CSV.Reader
   alias TdDd.Loader.LoaderWorker
   alias TdDd.Systems
   alias TdDd.Systems.System
 
-  @taxonomy_cache Application.get_env(:td_dd, :taxonomy_cache)
+  require Logger
 
   @structure_import_schema Application.get_env(:td_dd, :metadata)[:structure_import_schema]
   @structure_import_required Application.get_env(:td_dd, :metadata)[:structure_import_required]
@@ -24,12 +25,12 @@ defmodule TdDdWeb.MetadataController do
       do_upload(conn, params, system_id)
       send_resp(conn, :accepted, "")
     else
-      _ -> send_resp(conn, :not_found, Poison.encode!(%{error: "system.not_found"}))
+      _ -> send_resp(conn, :not_found, JSON.encode!(%{error: "system.not_found"}))
     end
   rescue
     e in RuntimeError ->
       Logger.error("While uploading #{e.message}")
-      send_resp(conn, :unprocessable_entity, Poison.encode!(%{error: e.message}))
+      send_resp(conn, :unprocessable_entity, JSON.encode!(%{error: e.message}))
   end
 
   @doc """
@@ -76,7 +77,7 @@ defmodule TdDdWeb.MetadataController do
   rescue
     e in RuntimeError ->
       Logger.error("While uploading #{e.message}")
-      send_resp(conn, :unprocessable_entity, Poison.encode!(%{error: e.message}))
+      send_resp(conn, :unprocessable_entity, JSON.encode!(%{error: e.message}))
   end
 
   defp do_upload(conn, params, system_id \\ nil) do
@@ -94,7 +95,7 @@ defmodule TdDdWeb.MetadataController do
   defp parse_data_structures(nil, _), do: {:ok, []}
 
   defp parse_data_structures(%Upload{path: path}, system_id) do
-    domain_map = @taxonomy_cache.get_domain_name_to_id_map()
+    domain_map = TaxonomyCache.get_domain_name_to_id_map()
     system_map = get_system_map(system_id)
 
     defaults =